fix: comprehensive security, concurrency, and UX fixes across entire codebase
Backend fixes (34 items):
- CRITICAL: Payment callback idempotency with pessimistic lock, null check, auto card generation
- CRITICAL: ApiKeyService @Transactional on all write methods, fix incrementUseCount race condition
- HIGH: JWT filter role whitelist validation, token secret length check (>=32 bytes)
- HIGH: Wallet consume pessimistic lock to prevent negative balance under concurrency
- HIGH: Password strength validation (>=6 chars) on change and reset flows
- HIGH: getUserInfo supports both admin and user roles
- HIGH: TOTP verification uses constant-time comparison to prevent timing attacks
- HIGH: adminAllOrders now respects status filter parameter
- MEDIUM: SecurityController pagination size cap (100), null IP handling
- MEDIUM: SettingsController dedup, jwt_secret masking
- MEDIUM: Card type mismatch fix (apiKeyId/mergedIntoCardId Long->Integer)
- MEDIUM: Entity updateTime auto-update via @UpdateTimestamp (ApiKey, Wallet, CardPricing)
- MEDIUM: DTO validations (RegisterBindRequest, LoginRequest @Size)
- MEDIUM: PaymentUtil order number collision risk reduced (8-digit random suffix)
- MEDIUM: CardPricingService @Transactional + existence check before delete
- MEDIUM: MaintenanceService NPE protection, BackupService error output separation
- MEDIUM: enableCard preserves previous status, new setCardStatus method
- MEDIUM: Phone uniqueness check in UserService.updateProfile

Frontend fixes (16 items):
- CRITICAL: UserPage v-if/v-else-if mixing fixed, userInfo sync with props via watch
- HIGH: SettingsPage payment_enabled boolean/string type consistency
- HIGH: api.js token refresh race condition (reset flag before processing queue)
- HIGH: OrdersManagePage pagination fix (handle Page object correctly)
- MEDIUM: UserPage usage records pagination (computed property)
- MEDIUM: OverviewPage chart empty data handling, high-DPI canvas support
- MEDIUM: NavigationBar admin modal pre-fill email
- MEDIUM: KeysManagePage export columns (status/create_time added to availableColumns)
- MEDIUM: UserManagePage exclude empty password on edit, toggleUserStatus confirmation
- MEDIUM: SettingsPage/NotificationPage remove window.location.reload()
- MEDIUM: PricingManagePage exclude null id on create
- MEDIUM: Dashboard remove orphaned carousel logic (dead code)
- MEDIUM: SettingsPage showToast use ElMessage instead of DOM manipulation
- LOW: loginform cancelOAuthRegister cleanup, App.vue maintenance interval optimization

+++ backend/src/main/java/org/xxg/backend/backend/dto/GenerateCardRequest.java
@@ -34,5 +34,5 @@ public class GenerateCardRequest {
     /** 有效期天数 */
     private Integer days;
     /** 关联的API密钥ID */
-    private Long apiKeyId;
+    private Integer apiKeyId;
 }
